Structure of STA side fix in place
Simply added some #ifdefs around definitions to ieee80211.h.
mlme.c now has the basic code structure in place to handle
deauthentication frames. The idea is to send out a deauth_verif
frame before notifying the cfg80211 layer of the deauthentication.
The appropriate information is saved away and a timer is set.
Once the timer fires, the cfg80211 layer is notified. of the
deauthentication. If a deauth_verif_resp is received, it deletes
the timer and life goes on as normal.